Product.java - Added comment

+++ src/main/java/se/lexicon/negar/jpaworkshopordermanagement/entity/Product.java
@@ -1,18 +1,18 @@
 package se.lexicon.negar.jpaworkshopordermanagement.entity;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 public class Product {
 
     private int Id;
     private String name;
-    private BigDecimal price;
+    // create as big decimal
+    private double price;
 
     public Product() {
     }
 
-    public Product(String name, BigDecimal price) {
+    public Product(String name, double price) {
         this.name = name;
         this.price = price;
     }
@@ -31,11 +31,11 @@ public class Product {
         this.name = name;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
